feat: update training configuration and dependencies

+++ trainer/trainer_utils.py
@@ -160,7 +160,7 @@ class SkipBatchSampler(Sampler):
 class LMForRewardModel:
     def __init__(self, model_path, device="cuda", dtype=torch.float16):
         self.tokenizer = AutoTokenizer.from_pretrained(model_path, trust_remote_code=True)
-        self.model = AutoModel.from_pretrained(model_path, torch_dtype=dtype, trust_remote_code=True)
+        self.model = AutoModel.from_pretrained(model_path, dtype=dtype, trust_remote_code=True)
         self.model = self.model.to(device).eval()
         self.device = device
